+++ sources/parsing.c
@@ -6,15 +6,15 @@
 /*   By: sbo <sbo@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/19 11:51:48 by sbo               #+#    #+#             */
-/*   Updated: 2024/04/22 12:18:08 by sbo              ###   ########.fr       */
+/*   Updated: 2024/04/22 13:13:57 by sbo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parsing.h"
 
-void	ft_memcpy(void *dst, void *src, size_t size)
+void	ft_memcpy(void *dst, void *src, int size)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < size)
@@ -26,7 +26,7 @@ void	ft_memcpy(void *dst, void *src, size_t size)
 
 bool	ft_str_match(char **str, char *prefix)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (prefix[i])
@@ -56,9 +56,9 @@ bool	ft_str_match_digit(char **str, uint8_t *digit)
 	return (false);
 }
 
-bool	ft_strnchr(char *str, size_t len, char c, size_t *pos)
+bool	ft_strnchr(char *str, int len, char c, int *pos)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < len)
